Make map() faster for Union{T, Missing/Nothing} element types
Storing typeof(el) in a variable S in the hot part of the loop kills
performance because S, which is inferred as equal to Union{Type{T}, Type{Missing/Nothing}},
is marked as ::Any. Only using it in the exceptional branch works around the problem
and makes map(identity, ::Vector{Union{Int, Missing}}) about 10 times faster.